Restructure codes
